Tested/fixed jdbc stuff, user stuff, etc.

+++ src/ie/dcu/redbrick/newsboards/server/database/NewsboardsDao.java
@@ -2,7 +2,7 @@ package ie.dcu.redbrick.newsboards.server.database;
 
 import ie.dcu.redbrick.newsboards.shared.Model;
 
-public abstract class NewsboardsDao <T extends Model<Integer>> extends BaseDao<T, Integer> {
+public abstract class NewsboardsDao <T extends Model<Long>> extends BaseDao<T, Long> {
     protected String getDsn() {
        return "java:comp/env/jdbc/newsgroupsdb";
     }
